# docs/swagger.yaml
definitions:
  body.EmailParam:
    properties:
      code:
        description: 验证码
        type: string
      email:
        description: 邮箱号
        type: string
    required:
    - code
    type: object
  body.MobileParam:
    properties:
      code:
        description: 验证码
        type: string
      mobile:
        description: 手机号
        type: string
    required:
    - code
    type: object
  body.UserCodeParam:
    properties:
      type:
        description: 找回密码find,注册reg,绑定bind
        type: string
      userLogin:
        description: 用户名
        type: string
    required:
    - userLogin
    type: object
  body.UserFindPassParam:
    properties:
      code:
        description: 验证码
        type: string
      user_login:
        description: 用户名
        type: string
      user_pass:
        description: 用户密码
        type: string
    required:
    - code
    - user_login
    - user_pass
    type: object
  body.UserLoginParam:
    properties:
      isAgent:
        type: integer
      userLogin:
        description: 用户名
        type: string
      userPass:
        description: 用户密码
        type: string
    required:
    - userLogin
    - userPass
    type: object
  body.UserLoginReturn:
    properties:
      id:
        type: integer
      token:
        type: string
    type: object
  body.UserRegisterParam:
    properties:
      hardwareId:
        description: 硬件ID
        type: string
      registerType:
        description: 1游客，默认0普通注册
        type: integer
      source:
        description: 注册来源
        type: string
      user_login:
        description: 游客类型不用传此参数
        type: string
      user_pass:
        description: 用户密码
        type: string
    required:
    - user_pass
    type: object
  http_error.HttpError:
    properties:
      code:
        example: 400
        type: integer
      msg:
        example: status bad request
        type: string
    type: object
  success.InfoData:
    properties:
      code:
        example: 0
        type: integer
      data:
        additionalProperties: true
        type: object
      msg:
        example: ok
        type: string
    type: object
host: test.pgc.api.yimisaas.com
info:
  contact: {}
  description: 直播&点播 API文档.
  title: API 文档
  version: "1.67"
paths:
  /login/getCode:
    post:
      consumes:
      - application/json
      description: 获取验证码
      parameters:
      - description: 账号和类型，type:find找回密码，reg注册
        in: body
        name: codeParam
        required: true
        schema:
          $ref: '#/definitions/body.UserCodeParam'
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/success.InfoData'
        "1010":
          description: ""
          schema:
            $ref: '#/definitions/http_error.HttpError'
        "1011":
          description: ""
          schema:
            $ref: '#/definitions/http_error.HttpError'
        "1012":
          description: ""
          schema:
            $ref: '#/definitions/http_error.HttpError'
        "1013":
          description: ""
          schema:
            $ref: '#/definitions/http_error.HttpError'
      summary: 获取验证码
      tags:
      - 登录相关
  /login/userFindPass:
    post:
      consumes:
      - application/json
      description: 找回密码
      parameters:
      - description: 账号和密码和验证码
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/body.UserFindPassParam'
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/body.UserLoginReturn'
        "1002":
          description: ""
          schema:
            $ref: '#/definitions/http_error.HttpError'
        "1003":
          description: ""
          schema:
            $ref: '#/definitions/http_error.HttpError'
        "1004":
          description: ""
          schema:
            $ref: '#/definitions/http_error.HttpError'
      summary: 找回密码
      tags:
      - 登录相关
  /login/userLogin:
    post:
      consumes:
      - application/json
      description: 登录
      parameters:
      - description: 账号和密码
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/body.UserLoginParam'
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/body.UserLoginReturn'
        "1002":
          description: ""
          schema:
            $ref: '#/definitions/http_error.HttpError'
        "1003":
          description: ""
          schema:
            $ref: '#/definitions/http_error.HttpError'
        "1004":
          description: ""
          schema:
            $ref: '#/definitions/http_error.HttpError'
      summary: 登录
      tags:
      - 登录相关
  /login/userReg:
    post:
      consumes:
      - application/json
      description: 注册
      parameters:
      - description: 账号和密码
        in: body
        name: registerParam
        required: true
        schema:
          $ref: '#/definitions/body.UserRegisterParam'
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/success.InfoData'
        "1005":
          description: ""
          schema:
            $ref: '#/definitions/http_error.HttpError'
        "1006":
          description: ""
          schema:
            $ref: '#/definitions/http_error.HttpError'
        "1007":
          description: ""
          schema:
            $ref: '#/definitions/http_error.HttpError'
        "1008":
          description: ""
          schema:
            $ref: '#/definitions/http_error.HttpError'
        "1010":
          description: ""
          schema:
            $ref: '#/definitions/http_error.HttpError'
        "1012":
          description: ""
          schema:
            $ref: '#/definitions/http_error.HttpError'
      summary: 注册
      tags:
      - 登录相关
  /user/bindEmail:
    post:
      consumes:
      - application/json
      description: 绑定邮箱
      parameters:
      - description: token
        in: header
        name: Auth
        required: true
        type: string
      - description: '-'
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/body.EmailParam'
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/success.InfoData'
      summary: 绑定邮箱
      tags:
      - 用户相关
  /user/bindMobile:
    post:
      consumes:
      - application/json
      description: 绑定手机
      parameters:
      - description: token
        in: header
        name: Auth
        required: true
        type: string
      - description: '-'
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/body.MobileParam'
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/success.InfoData'
      summary: 绑定手机
      tags:
      - 用户相关
swagger: "2.0"